l_08_cassandra_and_kafka/producer/requirements.txt
kafka-python==2.0.2
pandas==1.4.2